src/requirements.txt
annotated-types==0.5.0
anyio==3.7.1
CacheControl==0.13.1
cachetools==5.3.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.6
colorama==0.4.6
cryptography==41.0.2
exceptiongroup==1.1.2
faiss-cpu==1.7.4
fastapi==0.100.0
filelock==3.12.2
firebase-admin==6.2.0
fsspec==2023.6.0
google-api-core==2.11.1
google-api-python-client==2.95.0
google-auth==2.22.0
google-auth-httplib2==0.1.0
google-cloud-core==2.3.3
google-cloud-firestore==2.11.1
google-cloud-storage==2.10.0
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.59.1
grpcio==1.56.2
grpcio-status==1.56.2
h11==0.14.0
httplib2==0.22.0
huggingface-hub==0.16.4
idna==3.4
Jinja2==3.1.2
joblib==1.3.1
MarkupSafe==2.1.3
mpmath==1.3.0
msgpack==1.0.5
networkx==3.1
nltk==3.8.1
numpy==1.25.1
packaging==23.1
Pillow==10.0.0
proto-plus==1.22.3
protobuf==4.23.4
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.0.3
pydantic_core==2.3.0
PyJWT==2.8.0
pyparsing==3.1.0
PyYAML==6.0.1
regex==2023.6.3
requests==2.31.0
rsa==4.9
safetensors==0.3.1
scikit-learn==1.3.0
scipy==1.11.1
sentence-transformers==2.2.2
sentencepiece==0.1.99
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
sympy==1.12
threadpoolctl==3.2.0
tokenizers==0.13.3
torch==2.0.1
torchvision==0.15.2
tqdm==4.65.0
transformers==4.31.0
typing_extensions==4.7.1
uritemplate==4.1.1
urllib3==1.26.16
uvicorn==0.23.1
